feat(comment-crud): Finish enough comment crud in server to move to client

+++ Services/PostService.cs
@@ -8,48 +8,40 @@ namespace teachers_lounge_server.Services
     public class PostService
     {
         private static PostRepository repo => new PostRepository();
-        // TODO: add a function that turns a post from DB into a Post that includes the optional mostPopularComment field (of type Comment)
-        public static Task<List<Post>> GetPostsBySubjectIds(ObjectId[] subjectIds)
+
+        public static async Task<List<ExpandedPost>> GetPostsBySubjectIds(ObjectId[] subjectIds, int depth = 1)
         {
-            return GetPostsByFieldIn("subjectId", subjectIds);
+            var posts = await GetPostsByFieldIn("subjectId", subjectIds);
+            return await Expand(posts, depth);
         }
 
-        public static async Task<Post?> GetPostById(ObjectId postId)
+        public static async Task<ExpandedPost?> GetPostById(ObjectId postId, int depth = 1)
         {
-            List<Post> posts = await GetPostsByField("_id", postId);
-
-            return posts.Count != 1 ? null : posts[0];
+            var posts = await GetPostsByField("_id", postId);
+            return posts.Count != 1 ? null : await Expand(posts[0], depth);
         }
 
-        public static Task<List<Post>> GetPostsByField<TValue>(string fieldName, TValue value)
+        private static Task<List<Post>> GetPostsByField<TValue>(string fieldName, TValue value)
         {
             return repo.GetPostsByField(fieldName, value);
         }
 
-        public static Task<List<Post>> GetPostsByFieldIn<TValue>(string fieldName, TValue[] values)
+        private static Task<List<Post>> GetPostsByFieldIn<TValue>(string fieldName, TValue[] values)
         {
             return repo.GetPostsByFieldIn(fieldName, values);
         }
 
-        public static Task<List<Post>> GetPostsByMultipleFilters(IEnumerable<FilterDefinition<BsonDocument>> filterList)
+        private static Task<List<Post>> GetPostsByMultipleFilters(IEnumerable<FilterDefinition<BsonDocument>> filterList)
         {
             return repo.GetPostsByMultipleFilters(filterList);
-        }
-
-        public static async Task<ReplaceOneResult> UpsertPost(string userId, Post post)
-        {
-            return await repo.UpsertPost(post);
         }
 
         public static async Task<bool> CanUserDeletePost(string? userId, string postId)
         {
             if (!postId.IsObjectId())
-            {
                 return false;
-            }
 
-            var post = await GetPostById(ObjectId.Parse(postId));
-
+            var post = await GetPostById(ObjectId.Parse(postId), 0);
             return post != null && await CanUserDeletePost(userId, post);
         }
 
@@ -59,17 +51,12 @@ namespace teachers_lounge_server.Services
             var subject = await AssociationService.GetAssociationById(ObjectId.Parse(post.subjectId));
 
             if (user == null || subject == null)
-            {
                 return false;
-            }
 
             if (userId == post.authorId)
-            {
                 return true;
-            }
 
             bool isAdmin = await UserService.HasPermissions(userId, Role.Admin);
-
             bool doShareSchool = user.associatedSchools.Some(schoolId => subject.associatedSchools.Contains(schoolId));
 
             return isAdmin && doShareSchool;
@@ -80,16 +67,50 @@ namespace teachers_lounge_server.Services
             return post.authorId == userId;
         }
 
-        // TODO: fix the delete method here so it deletes all the comments as well
+        public static Task<ReplaceOneResult> UpsertPost(string userId, Post post)
+        {
+            return repo.UpsertPost(post);
+        }
+
         public static Task<UpdateResult> IncrementTotalChildrenCount(ObjectId postId, int delta)
         {
             return repo.IncremeantTotalCommentsCount(postId, delta);
         }
 
-
-        public static Task<bool> DeletePost(string postId)
+        public static async Task<bool> DeletePost(string postId)
         {
-            return repo.DeletePost(postId);
+            if (!postId.IsObjectId())
+                return false;
+
+            var parsedPostId = ObjectId.Parse(postId);
+            var results = await Task.WhenAll([
+                repo.DeletePost(postId),
+                CommentService.DeleteAllCommentsInPost(parsedPostId)
+            ]);
+
+            return results?.Every(result => result) ?? false;
+        }
+
+        private static async Task<ExpandedPost> Expand(Post post, int depth = 1)
+        {
+            try
+            {
+                var comments = await CommentService.GetCommentsByParentId(ObjectId.Parse(post.id), depth - 1);
+                return new ExpandedPost(post, comments.ToArray());
+            }
+            catch
+            {
+                return new ExpandedPost(post);
+            }
+        }
+
+        private static async Task<List<ExpandedPost>> Expand(IEnumerable<Post> posts, int depth = 1)
+        {
+            if (depth <= 0)
+                return posts.Map(p => new ExpandedPost(p));
+
+            var tasks = posts.Map(post => Expand(post, depth));
+            return (await Task.WhenAll(tasks)).ToList();
         }
     }
 }
